moved todos out of schema into own file
