+++ src/db/schema.ts
@@ -1,28 +1,11 @@
-import {
-  boolean,
-  integer,
-  pgTable,
-  timestamp,
-  varchar,
-  text,
-} from "drizzle-orm/pg-core"
 import { createSchemaFactory } from "drizzle-zod"
 import { z } from "@hono/zod-openapi"
 export * from "./auth-schema"
-import { users } from "./auth-schema"
+export * from "./app-schema"
+import { todosTable } from "./app-schema"
 
 const { createInsertSchema, createSelectSchema, createUpdateSchema } =
   createSchemaFactory({ zodInstance: z })
-
-export const todosTable = pgTable(`todos`, {
-  id: integer().primaryKey().generatedAlwaysAsIdentity(),
-  text: varchar({ length: 500 }).notNull(),
-  completed: boolean().notNull().default(false),
-  created_at: timestamp({ withTimezone: true }).notNull().defaultNow(),
-  user_id: text("user_id")
-    .notNull()
-    .references(() => users.id, { onDelete: "cascade" }),
-})
 
 export const selectTodoSchema = createSelectSchema(todosTable)
 export const createTodoSchema = createInsertSchema(todosTable)
